增加布局的初步测试

+++ WangShunManager/Dtos/UserRowDto.cs
@@ -6,9 +6,9 @@
         public string LoginId { get; set; }
         public string RealName { get; set; }
         public string CompanyName { get; set; }
-        public int UsableBalance { get; set; }
-        public int FreezeBalance { get; set; }
-        public int CreditAmount { get; set; }
+        public decimal UsableBalance { get; set; }
+        public decimal FreezeBalance { get; set; }
+        public decimal CreditAmount { get; set; }
         public string CompanyAddress { get; set; }
         public string ContactQq { get; set; }
         public string Email { get; set; }
